Fix calling 'OnPedExitedSphereEx' event unneccessary.
forEach loop scans for all created spheres, so when there are more than one sphere and we are not in "enter" distance to the any other found spheres, the OnPedExitedSphereEx event will be called. Checking 'sphere == tempSphere' will call exit event only when we are exiting the real entered sphere.